Relocated config.json to constants/ and refactored code associated with it
